# docs/swagger.yaml
definitions:
...
  forms.ArticleCategoryAddForm:
    properties:
      name:
        description: 分类名称
        type: string
      parent_id:
        description: 父级分类 ID
        type: integer
      type:
        description: 分类类型，默认值为 0 表文章；1 表友链
        type: integer
      url:
        description: 访问 URL
        type: string
    required:
    - name
    - url
    type: object
  forms.ArticleCategoryEditForm:
    properties:
      ID:
        type: integer
      name:
        description: 分类名称
        type: string
      parent_id:
        description: 父级分类 ID
        type: integer
      type:
        description: 分类类型，默认值为 0 表文章；1 表友链
        type: integer
      url:
        description: 访问 URL
        type: string
    required:
    - ID
    - name
    - url
    type: object
...
  forms.CommentAddForm:
    properties:
      admin_user_id:
        description: 博主 ID
        type: integer
      article_id:
        description: 文章 ID
        type: integer
      content:
        description: 评论内容
        type: string
      email:
        description: 邮箱
        type: string
      is_checked:
        description: 是否通过审核
        type: boolean
      is_recycled:
        description: 是否加入回收站
        type: boolean
      md_content:
        description: markdown 渲染后评论内容
        type: string
      nick_name:
        description: 昵称
        type: string
      parent_comment_id:
        description: 父评论 ID
        type: integer
      root_comment_id:
        description: 根评论 ID
        type: integer
      url:
        description: 访问地址
        type: string
    required:
    - article_id
    - content
    - email
    - nick_name
    - url
    type: object
  forms.CommentEditForm:
    properties:
      admin_user_id:
        description: 博主 ID
        type: integer
      article_id:
        description: 文章 ID
        type: integer
      content:
        description: 评论内容
        type: string
      email:
        description: 邮箱
        type: string
      id:
        description: ID
        type: integer
      is_checked:
        description: 是否通过审核
        type: boolean
      is_recycled:
        description: 是否加入回收站
        type: boolean
      md_content:
        description: markdown 渲染后评论内容
        type: string
      nick_name:
        description: 昵称
        type: string
      parent_comment_id:
        description: 父评论 ID
        type: integer
      root_comment_id:
        description: 根评论 ID
        type: integer
      url:
        description: 访问地址
        type: string
    required:
    - article_id
    - content
    - email
    - id
    - nick_name
    - url
    type: object
...
  forms.LinkCategoryAddForm:
    properties:
      name:
        description: 分类名称
        type: string
      type:
        description: 分类类型，默认值为 0 表文章；1 表友链
        type: integer
    required:
    - name
    type: object
  forms.LinkCategoryEditForm:
    properties:
      ID:
        type: integer
      name:
        description: 分类名称
        type: string
      type:
        description: 分类类型，默认值为 0 表文章；1 表友链
        type: integer
    required:
    - ID
    - name
    type: object
...
  utils.Result:
    properties:
      code:
        description: 状态码
        example: 100
        type: integer
      data:
        description: 数据
        type: object
      msg:
        description: 提示
        example: 密码错误
        type: string
    type: object
host: localhost:8088
info:
  contact: {}
  description: Aries 开源博客项目 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Swagger
  version: "1.0"
paths:
...
  /api/v1/all_categories:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分类类型
        in: query
        name: category_type
        required: true
        type: integer
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 获取所有分类
      tags:
      - 分类
  /api/v1/all_comments:
    get:
      consumes:
      - application/json
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 获取所有评论
      tags:
      - 评论
...
  /api/v1/categories:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ids
        in: query
        name: ids
        required: true
        type: string
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 批量删除分类
      tags:
      - 分类
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: size
        type: integer
      - description: 关键词
        in: query
        name: key
        type: string
      - description: 分类类型
        in: query
        name: category_type
        required: true
        type: integer
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 分页获取分类数据
      tags:
      - 分类
  /api/v1/categories/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 删除分类
      tags:
      - 分类
  /api/v1/categories/article:
    post:
      consumes:
      - application/json
      parameters:
      - description: 添加文章分类表单
        in: body
        name: addForm
        required: true
        schema:
          $ref: '#/definitions/forms.ArticleCategoryAddForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 添加文章分类
      tags:
      - 分类
    put:
      consumes:
      - application/json
      parameters:
      - description: 修改文章分类表单
        in: body
        name: editForm
        required: true
        schema:
          $ref: '#/definitions/forms.ArticleCategoryEditForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 修改文章分类
      tags:
      - 分类
  /api/v1/categories/link:
    post:
      consumes:
      - application/json
      parameters:
      - description: 添加友链分类表单
        in: body
        name: addForm
        required: true
        schema:
          $ref: '#/definitions/forms.LinkCategoryAddForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 添加友链分类
      tags:
      - 分类
    put:
      consumes:
      - application/json
      parameters:
      - description: 修改友链分类表单
        in: body
        name: editForm
        required: true
        schema:
          $ref: '#/definitions/forms.LinkCategoryEditForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 修改友链分类
      tags:
      - 分类
  /api/v1/comments:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ids
        in: query
        name: ids
        required: true
        type: string
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 批量删除评论
      tags:
      - 评论
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: size
        type: integer
      - description: 关键词
        in: query
        name: key
        type: string
      - description: 类型
        in: query
        name: type
        type: string
      - description: 状态
        in: query
        name: state
        type: integer
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 分页获取评论
      tags:
      - 评论
    post:
      consumes:
      - application/json
      parameters:
      - description: 添加评论表单
        in: body
        name: form
        schema:
          $ref: '#/definitions/forms.CommentAddForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 发表评论
      tags:
      - 评论
    put:
      consumes:
      - application/json
      parameters:
      - description: 修改评论表单
        in: body
        name: form
        schema:
          $ref: '#/definitions/forms.CommentEditForm'
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 修改评论
      tags:
      - 评论
  /api/v1/comments/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 删除评论
      tags:
      - 评论
...
  /api/v1/parent_categories:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分类类型
        in: query
        name: category_type
        required: true
        type: integer
      responses:
        "100":
          description: Continue
          schema:
            $ref: '#/definitions/utils.Result'
        "104":
          schema:
            $ref: '#/definitions/utils.Result'
      summary: 获取所有父级分类
      tags:
      - 分类
...
swagger: "2.0"